+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+0.7.2 (2020-06-27)
+------------------
+
+* Added option for disabling verification by HTTP request of default /favicon.ico file, to have similar behavior
+  to tag-based favicon results.
+* Fixed flaky test.
+
 0.7.1 (2020-06-17)
 ------------------